feat(scanner): Scanner now supports comment.
This resulted in modifying how the next character is read. Instead of having an increment of
the current column at the end of each loop, it is removed and it is done manually using the
advance method. This is to prevent double increments of the column variable resulting in
skipped characters.